Use async channel strategy for WSGI response iteration
