+++ src/wsgi/types.rs
@@ -1,16 +1,16 @@
 use futures::StreamExt;
 use http_body_util::BodyExt;
 use hyper::body::{self, Bytes};
-use pyo3::prelude::*;
-use pyo3::types::{PyBytes, PyList};
-use std::borrow::Cow;
+use pyo3::{
+    prelude::*,
+    types::{PyBytes, PyList},
+};
 use std::sync::{Arc, Mutex};
 use tokio::sync::Mutex as AsyncMutex;
 use tokio_util::bytes::{BufMut, BytesMut};
 
 use crate::conversion::BytesToPy;
 use crate::runtime::RuntimeRef;
-use crate::utils::log_application_callable_exception;
 
 const LINE_SPLIT: u8 = u8::from_be_bytes(*b"\n");
 
@@ -163,72 +163,5 @@ impl WSGIBody {
             .map(|item| PyBytes::new_bound(py, item))
             .collect();
         PyList::new_bound(py, lines)
-    }
-}
-
-pub(crate) struct WSGIResponseBodyIter {
-    inner: Option<PyObject>,
-    closed: bool,
-}
-
-impl WSGIResponseBodyIter {
-    pub fn new(body: PyObject) -> Self {
-        Self {
-            inner: Some(body),
-            closed: false,
-        }
-    }
-
-    #[inline]
-    fn close_inner(&mut self, py: Python) {
-        let _ = self
-            .inner
-            .as_ref()
-            .unwrap()
-            .call_method0(py, pyo3::intern!(py, "close"));
-        self.closed = true;
-    }
-}
-
-impl Drop for WSGIResponseBodyIter {
-    fn drop(&mut self) {
-        if !self.closed {
-            Python::with_gil(|py| self.close_inner(py));
-        }
-        let inner = self.inner.take().unwrap();
-        Python::with_gil(|_| drop(inner));
-    }
-}
-
-impl Iterator for WSGIResponseBodyIter {
-    type Item = Result<body::Frame<Bytes>, anyhow::Error>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        Python::with_gil(|py| {
-            match self
-                .inner
-                .as_ref()
-                .unwrap()
-                .call_method0(py, pyo3::intern!(py, "__next__"))
-            {
-                Ok(chunk_obj) => match chunk_obj.extract::<Cow<[u8]>>(py) {
-                    Ok(chunk) => {
-                        let chunk: Box<[u8]> = chunk.into();
-                        Some(Ok(body::Frame::data(Bytes::from(chunk))))
-                    }
-                    _ => {
-                        self.close_inner(py);
-                        None
-                    }
-                },
-                Err(err) => {
-                    if !err.is_instance_of::<pyo3::exceptions::PyStopIteration>(py) {
-                        log_application_callable_exception(&err);
-                    }
-                    self.close_inner(py);
-                    None
-                }
-            }
-        })
     }
 }
